feat(test): move the serializer separate of the test
